Made a resuable toolbar class and applied it to all major industry pages
