+++ team_b/yappy/lib/help.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:yappy/tool_bar.dart';
 
 
 class HelpApp extends StatelessWidget {
@@ -11,29 +12,16 @@ class HelpApp extends StatelessWidget {
 }
 
 class HelpPage extends StatelessWidget {
+  @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text('v'),
+      backgroundColor: const Color.fromARGB(255, 0, 0, 0),
+
+      appBar: PreferredSize(
+        preferredSize: Size.fromHeight(140), 
+        child: ToolBar()
       ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'Welcome to our Help Page!',
-              style: TextStyle(fontSize: 24),
-            ),
-            SizedBox(height: 20),
-            ElevatedButton(
-              onPressed: () {
-                // Add your onPressed code here!
-              },
-              child: Text('View Menu'),
-            ),
-          ],
-        ),
-      ),
+      drawer: HamburgerDrawer(),
     );
   }
 }
