+++ team_b/yappy/lib/medical_doctor.dart
@@ -1,40 +1,27 @@
 import 'package:flutter/material.dart';
+import 'package:yappy/tool_bar.dart';
 
 
-class Medical_doctorApp extends StatelessWidget {
+class Medical_DoctorApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Medical_doctorPage(),
+      home: Medical_DoctorPage(),
     );
   }
 }
 
-class Medical_doctorPage extends StatelessWidget {
+class Medical_DoctorPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text('Medical_doctor'),
+      backgroundColor: const Color.fromARGB(255, 0, 0, 0),
+
+      appBar: PreferredSize(
+        preferredSize: Size.fromHeight(140), 
+        child: ToolBar()
       ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'Welcome to our Medical_doctor App!',
-              style: TextStyle(fontSize: 24),
-            ),
-            SizedBox(height: 20),
-            ElevatedButton(
-              onPressed: () {
-                // Add your onPressed code here!
-              },
-              child: Text('View Menu'),
-            ),
-          ],
-        ),
-      ),
+      drawer: HamburgerDrawer(),
     );
   }
 }
