+++ team_b/yappy/lib/medical_patient.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-
+import 'package:yappy/tool_bar.dart';
 
 class Medical_PatientApp extends StatelessWidget {
   @override
@@ -14,27 +14,13 @@ class Medical_PatientPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text('Medical_Patient'),
+      backgroundColor: const Color.fromARGB(255, 0, 0, 0),
+
+      appBar: PreferredSize(
+        preferredSize: Size.fromHeight(140), 
+        child: ToolBar()
       ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'Welcome to our Medical_Patient APP!',
-              style: TextStyle(fontSize: 24),
-            ),
-            SizedBox(height: 20),
-            ElevatedButton(
-              onPressed: () {
-                // Add your onPressed code here!
-              },
-              child: Text('View Menu'),
-            ),
-          ],
-        ),
-      ),
+      drawer: HamburgerDrawer(),
     );
   }
 }
